Minor refactoring.

+++ py_misc_utils/select_params.py
@@ -163,7 +163,7 @@ class Selector:
   def _random_generate(self, count):
     rng = np.random.default_rng()
     low = np.zeros(len(self.space), dtype=np.int32)
-    high = np.array(self.space)
+    high = np.array(self.space, dtype=np.int32)
 
     rpoints = []
     for _ in range(count):
